Move shared logic for test into helper file
